refactor(inventario): crear modelos y mocks de conciliacion, migrar detalle a signals
