+++ libs/inventario/inventario/src/lib/pages/conciliacion-page/conciliacion-historial/conciliacion-historial.component.html
@@ -99,7 +99,7 @@
           </tr>
         </thead>
         <tbody>
-          @for (reg of registros; track reg.id) {
+          @for (reg of registros(); track reg.id) {
             <tr>
               <td class="font-bold text-primary">{{ reg.id }}</td>
               <td class="text-secondary whitespace-pre">{{ reg.fecha.replace(', ', '\n') }}</td>
